requirements.txt
rich~=10.2.1
click~=8.0.0


setuptools~=56.2.0
numpy~=1.20.3
torch~=1.8.1
opencv-python~=4.5.2.52
torchmetrics~=0.3.2
tifffile~=2021.4.8
PyYAML~=5.4.1
captum~=0.4.0
matplotlib~=3.4.2
tqdm~=4.60.0
seaborn~=0.11.1
